feat: Add `no_html` and `no_png` options to `rendercv_settings`
This commit adds the `no_html` and `no_png` options to the `rendercv_settings` block in the `John_Doe_ClassicTheme_CV.yaml` file. These options allow the user to control whether the HTML and PNG files will be generated during the rendering process. The default values are set to `true` for `no_html` and `false` for `no_png`.