refactor: Update rendercv_settings to allow disabling HTML and PNG generation
This commit updates the `rendercv_settings` in the `John_Doe_ClassicTheme_CV.yaml` file to allow disabling the generation of HTML and PNG files. The `no_html` and `no_png` flags are set to `true` in the `render` section of the settings. This change provides more flexibility to users who may not need these file formats in their rendered output.